+++ CHANGELOG.md
@@ -18,10 +18,12 @@ This change log adheres to standards from [Keep a CHANGELOG](https://keepachange
 * [Tests] add @typescript-eslint/parser v6 ([#3629][] @HenryBrown0)
 * [Tests] add @typescript-eslint/parser v7 and v8 ([#3629][] @hampustagerud)
 * [Docs] [`no-danger`]: update broken link ([#3817][] @lucasrmendonca)
+* [types] add jsdoc type annotations ([#3731][] @y-hsgw)
 
 [#3632]: https://github.com/jsx-eslint/eslint-plugin-react/issues/3632
 
 [#3812]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3812
+[#3731]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3731
 [#3629]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3629
 [#3817]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3817
 [#3807]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3807